requirements.txt
requests==2.31.0
python-dotenv==1.0.0
sqlite3==2.6.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2 